requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.5
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
httpie==2.3.0
idna==2.10
iniconfig==1.1.1
Markdown==3.3.3
packaging==20.8
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
Pygments==2.7.4
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.1
pytest-django==4.1.0
pytz==2020.5
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
